Connect dash hud and dash ability

+++ Assets/Script/Data/PlayerData.cs
@@ -12,7 +12,7 @@ namespace SGGames.Script.Data
         [Header("Dash Data")]
         [SerializeField] private int m_dashDistance;
         [SerializeField] private float m_dashSpeed;
-        [SerializeField] private float m_dashCooldown;
+        [SerializeField] private int m_dashCooldown;
         [SerializeField] private int m_staminaCostForDash;
         [Header("Resistance(%)")]
         [SerializeField][Range(0,100)] private float m_defaultDamageResistance;
@@ -28,7 +28,7 @@ namespace SGGames.Script.Data
         public int ActionPoints => m_actionPoints;
         public int DashDistance => m_dashDistance;
         public float DashSpeed => m_dashSpeed;
-        public float DashCooldown => m_dashCooldown;
+        public int DashCooldown => m_dashCooldown;
         public float MaxDashSpeed => m_maxDashSpeed;
         public int StaminaCostForDash => m_staminaCostForDash;
         public float DefaultDamageResistance => m_defaultDamageResistance;
